chore: move NATO/APCO logic to own file
